move some logic around. rename some things
