+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "shamos-hoey",
-  "version": "1.1.1",
+  "version": "1.1.0",
   "description": "A module to check if a polygon self-intersects",
   "main": "dist/shamosHoey.js",
   "module": "dist/shamosHoey.esm.js",
