+++ src/main.js
@@ -1,36 +1,10 @@
 import EventQueue from './EventQueue'
-import Sweepline from './Sweepline'
+import runCheck from './runCheck'
 
-import {fillEventQueue} from './fillQueue'
+import fillEventQueue from './fillQueue'
 
-// import {debugEventAndSegments, debugEventAndSegment} from './debug'
-
-export default function isSimple (geojson) {
-
+export default function noIntersections (geojson) {
     const eventQueue = new EventQueue()
-
     fillEventQueue(geojson, eventQueue)
-
-    const sweepLine = new Sweepline();
-
-    let currentSegment = null
-
-    while (eventQueue.length) {
-        const event = eventQueue.pop();
-
-        // debugEventAndSegments(event, sweepLine)
-
-        if (event.isLeftEndpoint) {
-            currentSegment = sweepLine.addSegment(event)
-            // debugEventAndSegment(event, currentSegment)
-            if (sweepLine.testIntersect(currentSegment, currentSegment.segmentAbove)) return false
-            if (sweepLine.testIntersect(currentSegment, currentSegment.segmentBelow)) return false
-        } else {
-            if (!event.segment) continue
-            // debugEventAndSegment(event, event.segment)
-            if (sweepLine.testIntersect(event.segment.segmentAbove, event.segment.segmentBelow)) return false
-            sweepLine.removeSegmentFromSweepline(event.segment)
-        }
-    }
-    return true
+    return runCheck(eventQueue)
 }
