+++ test/test.spec.js
@@ -4,7 +4,7 @@ import glob from 'glob'
 
 import load from 'load-json-file'
 
-import isSimple from '../src/main.js'
+import noIntersections from '../src/main.js'
 
 const trueFixtures = glob.sync(path.join(__dirname, 'fixtures', 'simple', '*.geojson'))
 const falseFixtures = glob.sync(path.join(__dirname, 'fixtures', 'notSimple', '*.geojson'))
@@ -13,7 +13,7 @@ test('simple fixtures', (t) => {
     trueFixtures.forEach((filepath) => {
         const name = path.parse(filepath).name;
         const geojson = load.sync(filepath);
-        t.true(isSimple(geojson), `[true] ${name}`);
+        t.true(noIntersections(geojson), `[true] ${name}`);
     });
 })
 
@@ -21,13 +21,13 @@ test('complex fixtures', (t) => {
     falseFixtures.forEach((filepath) => {
         const name = path.parse(filepath).name;
         const geojson = load.sync(filepath);
-        t.false(isSimple(geojson), `[false] ${name}`);
+        t.false(noIntersections(geojson), `[false] ${name}`);
     });
 })
 
 test('input data is not modified', (t) => {
     const geojson = load.sync(path.join(__dirname, 'fixtures', 'notSimple', 'switzerlandKinked.geojson'));
     const clonedData = JSON.parse(JSON.stringify(geojson))
-    isSimple(geojson)
+    noIntersections(geojson)
     t.deepEqual(geojson, clonedData)
 })
